weiss_core::legal

Enum ActionDesc

pub enum ActionDesc {

    MulliganConfirm,
    MulliganSelect {
        hand_index: u8,
    },
    Pass,
    Clock {
        hand_index: u8,
    },
    MainPlayCharacter {
        hand_index: u8,
        stage_slot: u8,
    },
    MainPlayEvent {
        hand_index: u8,
    },
    MainMove {
        from_slot: u8,
        to_slot: u8,
    },
    MainActivateAbility {
        slot: u8,
        ability_index: u8,
    },
    ClimaxPlay {
        hand_index: u8,
    },
    Attack {
        slot: u8,
        attack_type: AttackType,
    },
    CounterPlay {
        hand_index: u8,
    },
    LevelUp {
        index: u8,
    },
    EncorePay {
        slot: u8,
    },
    EncoreDecline {
        slot: u8,
    },
    TriggerOrder {
        index: u8,
    },
    ChoiceSelect {
        index: u8,
    },
    ChoicePrevPage,
    ChoiceNextPage,
    Concede,

}

Canonical action descriptor used as the truth representation of legal actions.

Variants

MulliganConfirm

Confirm mulligan without selecting additional cards.

MulliganSelect

Select a hand card for mulligan.

Fields

hand_index: u8

Pass

Pass the current decision.

Clock

Clock a hand card.

Fields

hand_index: u8

MainPlayCharacter

Play a character from hand to a stage slot.

Fields

hand_index: u8

stage_slot: u8

MainPlayEvent

Play an event from hand.

Fields

hand_index: u8

MainMove

Move a character between stage slots.

Fields

from_slot: u8

to_slot: u8

MainActivateAbility

Activate a character ability from a stage slot.

Fields

slot: u8

ability_index: u8

ClimaxPlay

Play a climax from hand.

Fields

hand_index: u8

Attack

Declare an attack from a stage slot.

Fields

slot: u8

attack_type: AttackType

CounterPlay

Play a counter from hand.

Fields

hand_index: u8

LevelUp

Select a card for level up.

Fields

index: u8

EncorePay

Pay encore for a character.

Fields

slot: u8

EncoreDecline

Decline encore for a character.

Fields

slot: u8

TriggerOrder

Select trigger order index.

Fields

index: u8

ChoiceSelect

Select a choice option by index.

Fields

index: u8

ChoicePrevPage

Page to previous choice options.

ChoiceNextPage

Page to next choice options.

Concede

Concede the game.